Execute all Vulkan tests if feature flag is missing
Prior to this change, an exception would be thrown if the
android.software.vulkan.deqp.level was not present for a device that
supports Vulkan, when considering whether or not a Vulkan test needs
to be executed.  This change tolerates the feature flag not being
available, but in that case forces every Vulkan test to execute.

Bug: 151115443
Change-Id: I0b152089a9f9fa31e06a8ee51b333cbb8394d82b