Disable tests failing on Cuttlefish
To support using ANGLE as the default OpenGL ES driver on Cuttlefish, we
need to enable ATP with SWANGLE on CF for continuous testing.

This CL disables additional tests that are known to fail, while those
failures are investigated to keep the runs green and make new failures
more obvious.

Bug: 169241278
Test: CtsDeqpTestCases
Change-Id: Ic198a5e7bcae530d42cf56a3368118cb66ecaec3